requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.4.2
python-dotenv>=1.0.0
pytest>=7.4.3
httpx>=0.25.1
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
redis>=5.0.1
psycopg2-binary>=2.9.9
validators>=0.22.0
pytest-asyncio>=0.23.2 